requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.0.6
djangorestframework==3.11.0
Flask==1.1.2
get==2019.4.13
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
MyApplication==0.1.0
post==2019.4.13
psycopg2-binary==2.8.5
public==2019.4.13
pytz==2020.1
query-string==2019.4.13
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
